// src/data/model/EquipmentSlot.ts
export enum EquipmentSlot {
    upper = "upper", middle = "middle", lower = "lower", armor = "armor", rWeapon = "rWeapon", lWeapon = "lWeapon", garment = "garment", shoes = "shoes", rAccessery = "rAccessery", lAccessery = "lAccessery",

    upperCostume = "upperCostume", middleCostume = "middleCostume", lowerCostume = "lowerCostume", armorShadow = "armorShadow", weaponShadow = "weaponShadow", sheidShadow = "sheidShadow", garmentCostume = "garmentCostume", shoesShadow = "shoesShadow", earringShadow = "earringShadow", pendantShadow = "pendantShadow",
}